Rename manipulator files for more clarity
