+++ src/manipulators/no_implicit_returns_manipulator.ts
