+++ src/manipulators/strict_null_checks_manipulator.ts
