+++ src/manipulators/strict_property_initialization_manipulator.ts
